Big refactoring III: move modules around
- filetransfer and transit where moved to the top level, blocking was moved *into* the top level (lib.rs).
- io moved into core
- Deleted some old tokio code
