--- examples/ws-core.rs
@@ -1,161 +1,0 @@
-use magic_wormhole::core::{
-    APIAction, APIEvent, Action, AnswerType, IOAction, IOEvent,
-    PeerMessage, WSHandle, WormholeCore,
-};
-use std::cell::RefCell;
-use std::rc::Rc;
-use url::Url;
-
-// Can ws do hostname lookup? Use ip addr, not localhost, for now
-const MAILBOX_SERVER: &str = "ws://relay.magic-wormhole.io:4000/v1";
-const APPID: &str = "lothar.com/wormhole/text-or-file-xfer";
-
-struct MyFactory {
-    wsh: WSHandle,
-    wcr: Rc<RefCell<WormholeCore>>,
-}
-
-struct MyHandler {
-    wsh: WSHandle,
-    wcr: Rc<RefCell<WormholeCore>>,
-    out: ws::Sender,
-}
-
-fn main() {
-    println!("start");
-    // for now, pretend that the WormholeCore is only ever going to ask us to
-    // make a single connection. Eventually, it will manage reconnects too,
-    // and we must be prepared to make multiple connections when it asks.
-
-    let mut wc = WormholeCore::new(APPID, MAILBOX_SERVER);
-    let wsh;
-    let ws_url;
-    let mut actions = wc.start();
-    if let Action::IO(IOAction::WebSocketOpen(handle, url)) =
-        actions.pop().unwrap()
-    {
-        wsh = handle;
-        ws_url = Url::parse(&url).unwrap();
-    } else {
-        panic!();
-    }
-
-    let f = MyFactory {
-        wsh,
-        wcr: Rc::new(RefCell::new(wc)),
-    };
-
-    // connect() blocks until disconnect, so the Manager drives everything
-    // connect() requires a closure, use Builder to provide a real Factory
-    //ws::connect(ws_url, f).unwrap();
-
-    let b = ws::Builder::new();
-    let mut w1 = b.build(f).unwrap();
-    w1.connect(ws_url).unwrap();
-    w1.run().unwrap();
-}
-
-impl ws::Factory for MyFactory {
-    type Handler = MyHandler;
-    fn connection_made(&mut self, out: ws::Sender) -> MyHandler {
-        MyHandler {
-            wsh: self.wsh,
-            wcr: Rc::clone(&self.wcr),
-            out,
-        }
-    }
-}
-
-impl ws::Handler for MyHandler {
-    fn on_open(&mut self, _: ws::Handshake) -> Result<(), ws::Error> {
-        println!("on_open");
-        let mut wc = self.wcr.borrow_mut();
-        let actions = wc.do_io(IOEvent::WebSocketConnectionMade(self.wsh));
-        process_actions(&self.out, actions);
-        // TODO: this should go just after .start()
-        let actions = wc.do_api(APIEvent::AllocateCode(2));
-        process_actions(&self.out, actions);
-        let offer = PeerMessage::new_offer_message("hello from rust!").serialize();
-        // let offer = json!({"offer": {"message": "hello from rust"}});
-        // then expect {"answer": {"message_ack": "ok"}}
-        let actions = wc.do_api(APIEvent::Send(offer.into_bytes()));
-        process_actions(&self.out, actions);
-
-        Ok(())
-    }
-    fn on_message(&mut self, msg: ws::Message) -> Result<(), ws::Error> {
-        println!("got message {}", msg);
-        let mut wc = self.wcr.borrow_mut();
-        let text = msg.as_text()?.to_string();
-        let rx = IOEvent::WebSocketMessageReceived(self.wsh, text);
-        let actions = wc.do_io(rx);
-        process_actions(&self.out, actions);
-        Ok(())
-    }
-    fn on_close(&mut self, code: ws::CloseCode, reason: &str) {
-        println!("closing {:?} {}", code, reason);
-        let mut wc = self.wcr.borrow_mut();
-        let actions = wc.do_io(IOEvent::WebSocketConnectionLost(self.wsh));
-        process_actions(&self.out, actions);
-    }
-}
-
-fn process_actions(out: &ws::Sender, actions: Vec<Action>) {
-    for a in actions {
-        match a {
-            Action::IO(io) => match io {
-                IOAction::WebSocketSendMessage(_wsh, msg) => {
-                    println!("sending {:?}", msg);
-                    out.send(msg).unwrap();
-                }
-                IOAction::WebSocketClose(_wsh) => {
-                    out.close(ws::CloseCode::Normal).unwrap();
-                }
-                // TODO: open new connections, handle timers
-                _ => {
-                    println!("action {:?}", io);
-                }
-            },
-            Action::API(api) => match api {
-                APIAction::GotMessage(msg) => {
-                    let message = String::from_utf8(msg).unwrap();
-                    let peer_msg = PeerMessage::deserialize(&message);
-                    match peer_msg {
-                        PeerMessage::Offer(_) => {
-                            panic!("Sender can't get offer")
-                        }
-                        PeerMessage::Error(errmsg) => println!(
-                            "remote error, transfer abandoned: {}",
-                            errmsg
-                        ),
-                        PeerMessage::Answer(answer) => match answer {
-                            AnswerType::MessageAck(ack) => {
-                                if ack == "ok" {
-                                    println!(
-                                        "Message transferred successfully!"
-                                    );
-                                } else {
-                                    println!(
-                                        "Failed to transfer message: {}",
-                                        ack
-                                    );
-                                }
-                            }
-                            AnswerType::FileAck(ack) => {
-                                if ack == "ok" {
-                                    println!("File transferred successfully");
-                                } else {
-                                    println!("File transfer failed: {}", ack);
-                                }
-                            }
-                        },
-                        PeerMessage::Transit(..) => {
-                            println!("transit message recieved")
-                        }
-                    }
-                }
-                _ => println!("action {:?}", api),
-            },
-        }
-    }
-}
